requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
paho-mqtt>=2.1,<3.0
pytest>=8.0,<9.0
